requirements.txt
langchain
numpy
streamlit
langchain-community
streamlit
PyPDF2
unstructured
nltk
sentence-transformers
langchain-huggingface
faiss-cpu
python-docx



## groq
langchain_groq
langchain_core
langdetect